refactor: extract BaseToolkitTheme for shared override DPs
Pull duplicated FontOverrideSource, ColorOverrideSource,
FontOverrideDictionary, ColorOverrideDictionary, Colors DPs and their
change handlers into an abstract BaseToolkitTheme in the shared
Uno.Toolkit.UI project. MaterialToolkitTheme and SimpleToolkitTheme
now extend it and only contain UpdateSource() with theme-specific logic.
